WIP try out separate physics thread - aim to avoid blocking non physics movement due to phy
FIX refactoring snapshot registry
- need to setup logic for both physics and view thread

FIX sync nodes via snapshots
- i.e. m_allNodes is ^*out-of-sync*

FIX reduce locking

FIX improve sync

WIP copy object snapshots

FIX minor optimization in update

+++ src/particle/ParticleGenerator.cpp
@@ -55,7 +55,7 @@ namespace particle {
             particle.m_scale = 0.0001f + prnd(0.2f);
             particle.m_materialIndex = m_material.m_registeredIndex;
 
-            const int idx = floor(prnd(m_material.spriteCount));
+            const int idx = static_cast<int>(floor(prnd(m_material.spriteCount)));
 
             particle.m_spriteIndex = idx;
             particle.m_spriteCount = m_material.spriteCount;
